Update olive_es.ts
I am a perfectionist and I observed that some words remained untranslated so I looked for lines of several ts files until I got a 100% translation in Spanish. I think that this ts file is the definitive one to have the Alpha version totally in Spanish.
The reason to change "es_ES" to "es" is because this translation is valid for any Spanish-speaking country and not only for Spain.

A pleasure to collaborate with this project and excuse my lack of programming knowledge, I work in the audiovisual field.