Don't use GNU grep for OpenBSD package builds
Hi --

The -P flag is only available on GNU grep, not BSD grep. This causes package builds on OpenBSD to needlessly fail when running cmake.

We will only ever package tagged versions of Olive, so we'll all know what version of Olive we're using. If a user wants to build from the tip of the tree, then they'll have a .git directory and the fallback won't even happen.