Inroduce Node Styles
Short term solution to adding granularity
on node colouring and styles. Due to the
nature of QGraphics, the standard CSS
stylesheets are not available. This branch
adds in a cloned factory to introduce a
single object instance of each node style
that is reused amongst the nodes. It
uses Qt's STATIC_GLOBAL macro, so
should be thread safe.