Add Waveform Channel Swizzle
Adds four checkboxes to the waveform that select which channels are visable,
either R, G, B or Luma.

Ctrl clicking any channel displays that channel and turns off all the others.

Changes also allowed for a few optimizations to the waveform fragment shader as
now only the channels that are being displayed are shown. Also some values are
now precomputed.