Add Reset buttom to Preferences Panel
Add a SetValuesFromConfig(Config config) to each preferecnes tab.
This function allows us to set all our values based on a given config.
It is used to set the intial settings in the constructor and is called
when reseting to the default settings with a new config.

Due to the slightly different nature of how the audio tab already
handles its default settings the SetValuesFromConfig function makes
some assumptions. I.E. it assumes that if the config file value is ""
then to find the row that corresponds to the text "Default". Also added
a config map value for the recording format. Sample rate appears to be
still under development so has not been set up.

The keyboard tab already had a reset button. This is removed and
replaced with the new standard button. Functionality has been copied
over. As the keyboard shortcuts are not stored in the config this tab
has no SetValuesFromConfig() method.

All tabs now double check if you want to reset the values. Ctrl
clicking to reset all also generates a confirmation dialog.