Add UI controls to scopes.
Add a stacked widget that holds UI controls for the selected scope.

ScopeBase now has a pointer to a widget that can hold any UI controls
that a scope wants to set up. This keeps the relevant code for each
scope in its class as opposed to duming it all in the scope panel. It
also reduces the amount of connections that have to be made between
the controls and the scopes.

This is implemented for the waveform.