Decoder: ProbeMedia can deal with missing files
Previously if ProbeMedia was given a missing file it returned nullptr,
however this doesn't work when importing OTIO files that need their
footage to be relinked. What we do here instead is if the file can't be
found we create a blank FootagePtr and add the files basic metadata but
don't set the footage as valid. This then causes the Relink Media Dialog
to be run after the import.

This doesn't appear to cause any issues when importing footage fiels
normaly.