Add mutex around threaded calls to OCIO config
By design OCIO can hit assert errors when using the same OCIO::Config
instance in multiiple threads. Here we add a mutex to the ColorManager
and access methods so locks can be added around its use in threads.
This does not make the ColorManager automatically thread safe.

This commit also gives ColorProcessors access to the config so they can
lock it when required.

This covers the OCIO calls that were causing problems issue #1382,
however it is not impossible there are others that have been missed.