Fix #1708
When a node edge is first created its from_item and to_item are not
initilised. This was leading to the arrow at the end of the edge being
drawn in a random place. The work around here is to explicitly set the
arrow to be drawn only when the mouse has started moving and we know
where to draw it.