otioload: Add simple dialog to change settings
Adds a table of sequences in the OTIO file and allows the settings to
be changed. The dialog takes a list of sequences and relies on each
OTIO timeline having a unique name.

The dialog has to be called from the main thread so there is a wrapper
function in Core.