Allow CMake to generate compile_commands.json automatically in the build dir
Some editors rely on clangd for autocompletion. In those cases clangd
will look for a compile_commands.json file in the root of the repository
to obtain the list of flags used for compilation and the INCLUDEPATH for
each file.
This commit allows CMake to automatically generate such file in the
designated build folder.
To enable clangd to use that file, simply link it in the root of the
project repository.
The link will be ignored by git.